services/api-service/requirements.txt
# Core API Service Dependencies

# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Kafka
aiokafka==0.13.0      # updated from 0.9.1
kafka-python==2.0.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.5
pyjwt==2.8.0

# Validation & Serialization
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# HTTP Client
httpx==0.27.0        # updated from 0.26.0
aiohttp==3.9.3

# Utilities
python-dotenv==1.0.1
tenacity==8.2.3
structlog==24.1.0
orjson==3.9.13
ulid-py==1.1.0

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
factory-boy==3.3.0